server/requirements.txt
blinker==1.7.0
click==8.1.7
flask==3.0.0
itsdangerous==2.1.2
jinja2==3.1.2
markupsafe==2.1.3
werkzeug==3.0.1
psycopg2-binary